Create utils directory and refactoring path
