+++ src/mock/comment.js
@@ -1,5 +1,5 @@
 import {COMMENTS_COUNT} from "../const";
-import {getRandomArrayItem} from "../utils";
+import {getRandomArrayItem} from "../utils/common";
 
 const emotions = [
   `angry.png`,
