+++ src/utils/common.js
@@ -1,27 +1,4 @@
-import {EXTRA_MOVIE_COUNT, MONTH_NAMES} from "./const";
-
-const RenderPosition = {
-  AFTER_BEGIN: `afterbegin`,
-  BEFORE_END: `beforeend`
-};
-
-const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTER_BEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFORE_END:
-      container.append(element);
-      break;
-  }
-};
-
-const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
+import {EXTRA_MOVIE_COUNT, MONTH_NAMES} from "../const";
 
 const getTopRatingMovies = (movies) => {
   return movies.sort((a, b) => {
@@ -106,4 +83,4 @@ const getUserRankTitle = (rank) => {
   return rankTitle;
 };
 
-export {RenderPosition, render, createElement, getTopRatingMovies, randomInteger, getRandomArrayItem, getCommentTitles, formatFilmDuration, getMostCommentedFilms, convertsArrayToString, getCurrentComment, formatReleaseDate, formatCommentDate, getUserRankTitle};
+export {getTopRatingMovies, randomInteger, getRandomArrayItem, getCommentTitles, formatFilmDuration, getMostCommentedFilms, convertsArrayToString, getCurrentComment, formatReleaseDate, formatCommentDate, getUserRankTitle};
